pyproject.toml
[project]
name = "mcpdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.16",
    "mcp-use>=1.7.0",
    "python-dotenv>=1.0.1",
]
